pytest/requirements.txt
attrs==20.3.0
iniconfig==1.1.1
numpy==1.19.4
packaging==20.4
Pillow==8.0.1
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
pytest-html==3.0.0
pytest-metadata==1.10.0
selenium==3.141.0
six==1.15.0
toml==0.10.2
urllib3==1.25.11
